Move Java primitive extension package and add tests
