+++ src/main/resources/templates/bootstrap/input-textarea.mustache
@@ -6,8 +6,7 @@
               {{#placeholder}}placeholder="{{.}}"{{/placeholder}}
               {{#rows}}rows="{{.}}"{{/rows}}
               {{#required}}required{{/required}}
-              class="form-control {{#has_error}}is-invalid{{/has_error}}">
-    </textarea>
+              class="form-control {{#has_error}}is-invalid{{/has_error}}"></textarea>
 
     {{#has_single_error}}
         <div class="invalid-feedback">
