+++ src/test/scala/com/hhandoko/aktiform/app/view/render/bootstrap/InputTextComponentSpec.scala
@@ -7,8 +7,8 @@ import org.junit.runner.RunWith
 import org.scalatest.{MustMatchers, WordSpec}
 import org.scalatestplus.junit.JUnitRunner
 
+import com.hhandoko.aktiform.RandomFixture
 import com.hhandoko.aktiform.api.html.input.{FormFieldError, InputTextField}
-import com.hhandoko.aktiform.app.RandomFixture
 
 @RunWith(classOf[JUnitRunner])
 class InputTextComponentSpec extends WordSpec with MustMatchers with RandomFixture {
